+++ workflow/scripts/downsample_matrices.py
@@ -4,7 +4,7 @@ from cfdna.preprocessing.matrices import downsample_matrix
 
 if 'snakemake' in globals():
     with open(snakemake.input['mincov']) as f:
-        hardcut = int(f.read().strip())
+        hardcut = float(f.read().strip())
 
     matrix = np.load(snakemake.input['raw'])
     downsampled = downsample_matrix(matrix, hardcut)
